Move cluster animations from tests/visual_inspection/ to docs/notebooks/
PR #181 review (teutoburg) preferred that the cluster GIFs and their
producing scripts live with the docs, not under tests/. The
animations are demos, not test assertions - tests/ runs in CI without
visual review, whereas docs/notebooks/ feeds RTD where the GIFs can
be seen.

- Move plot_galactic_centre_cluster.py -> docs/notebooks/plot_galactic_centre.py
- Move plot_globular_cluster.py -> docs/notebooks/plot_globular_cluster.py
- Move both GIFs to docs/notebooks/ next to their scripts.
- Delete the now-empty scopesim_templates/tests/visual_inspection/.
- Add two MyST notebooks (galactic_centre.md, globular_cluster.md)
  with a brief description, a quick `src.fields[0].field` example,
  and the pre-rendered GIF embedded via `:figure:`. The GIFs are not
  regenerated on doc-build (matplotlib FuncAnimation is too slow);
  contributors re-run the scripts manually to refresh them.
- Wire the two new notebooks into docs/index.md toctree.
