+++ docs/notebooks/galactic_centre_evolution.gif
