+++ docs/notebooks/globular_cluster_evolution.gif
